# api-svc/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Jewelry API",
        "title": "Jewelry API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users": {
            "get": {
                "description": "returns all products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "get all products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/services.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/managers.ProductView"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/services.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/services.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/services.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "creates order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "create order",
                "parameters": [
                    {
                        "description": "create order request",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateOrder"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/services.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/managers.OrderView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/services.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "obj0ect"
                        }
                    }
                }
            }
        },
        "/users/{status}": {
            "get": {
                "description": "returns products by status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "get products by status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product status",
                        "name": "status",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/services.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/managers.ProductView"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/services.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/services.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/services.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "managers.OrderView": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/types.Order"
                },
                "productView": {
                    "$ref": "#/definitions/managers.ProductView"
                }
            }
        },
        "managers.ProductView": {
            "type": "object",
            "properties": {
                "material": {
                    "$ref": "#/definitions/types.Material"
                },
                "product": {
                    "$ref": "#/definitions/types.Product"
                },
                "status": {
                    "$ref": "#/definitions/types.Status"
                }
            }
        },
        "services.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "types.CreateOrder": {
            "type": "object",
            "properties": {
                "delivery": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "order_email": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
...
        "types.Material": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "types.Order": {
            "type": "object",
            "properties": {
                "delivery": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "manager_id": {
                    "type": "string"
                },
                "order_email": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "types.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "material_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "status_id": {
                    "$ref": "#/definitions/types.StatusID"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "types.Status": {
            "type": "object",
            "properties": {
                "id": {
                    "$ref": "#/definitions/types.StatusID"
                },
                "text_code": {
                    "$ref": "#/definitions/types.StatusTextCode"
                }
            }
        },
        "types.StatusID": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "LockedInt",
                "UnLockedInt",
                "SoldInt"
            ]
        },
        "types.StatusTextCode": {
            "type": "string",
            "enum": [
                "LOCKED",
                "UNLOCKED",
                "SOLD"
            ],
            "x-enum-varnames": [
                "Locked",
                "UnLocked",
                "Sold"
            ]
        },
...
    }
}